# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "xinfra 平台后端 API 文档",
        "title": "xinfra API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/test/error/{code}": {
            "get": {
                "description": "根据错误码返回对应的业务错误响应",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试业务错误响应",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "业务错误码 (10001-10099: 认证相关, 20001-20099: 任务相关)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "details": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/test/timeout": {
            "get": {
                "description": "模拟请求超时（15秒延迟，超过前端 10s timeout）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试超时响应",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0 表示成功",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据，成功时返回"
                },
                "details": {
                    "description": "错误详情，失败时返回",
                    "type": "string"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 Bearer Token（例如：Bearer xxx）",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}